Ajout d'un script permettant de lancer le projet sans passer par un ide.

+++ Projet_1-Calculateur_de_notes/Note_Calculator.Tests/Student_MentionTest.cs
@@ -6,12 +6,12 @@ using Note_Calculator.Domain;
 public class Students_withSpecificAverage_GetSpecificMention
 {
     [Theory]
-    [InlineData(new int[] { 8, 9, 10 }, "Refusé")]
-    [InlineData(new int[] { 10, 10, 11 }, "Passable")]
-    [InlineData(new int[] { 12, 13, 14 }, "Assez bien")]
-    [InlineData(new int[] { 14, 15, 16 }, "Bien")]
-    [InlineData(new int[] { 17, 18, 19 }, "Très bien")]
-    public void studentIsValidatingYear(int[] notes, string expectedMention)
+    [InlineData(new double[] { 8, 9, 10 }, "Refusé")]
+    [InlineData(new double[] { 10, 10, 11 }, "Passable")]
+    [InlineData(new double[] { 12, 13, 14 }, "Assez bien")]
+    [InlineData(new double[] { 14, 15, 16 }, "Bien")]
+    [InlineData(new double[] { 17, 18, 19 }, "Très bien")]
+    public void studentIsValidatingYear(double[] notes, string expectedMention)
     {
         //Arrange:  On donne le contexte du test
         var student = new Student();
